+++ src/main/java/io/electrum/prepaidutility/api/KeyChangeTokenRequestsResource.java
@@ -15,8 +15,10 @@ import javax.ws.rs.core.SecurityContext;
 import javax.ws.rs.core.UriInfo;
 
 import io.electrum.prepaidutility.model.ErrorDetail;
+import io.electrum.prepaidutility.model.KeyChangeConfirmation;
 import io.electrum.prepaidutility.model.KeyChangeTokenRequest;
 import io.electrum.prepaidutility.model.KeyChangeTokenResponse;
+import io.electrum.vas.model.BasicAdvice;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
@@ -40,6 +42,19 @@ public abstract class KeyChangeTokenRequestsResource {
 
       public class PathParameters {
          public static final String REQUEST_ID = "requestId";
+      }
+   }
+
+   public class ConfirmKeyChange {
+      public static final String CONFIRM_KEY_CHANGE = "confirmKeyChange";
+      public static final int SUCCESS = 202;
+      public static final String RELATIVE_PATH =
+            "/{" + PathParameters.REQUEST_ID + "}/confirmations/{" + PathParameters.CONFIRMATION_ID + "}";
+      public static final String PATH = KeyChangeTokenRequestsResource.PATH + RELATIVE_PATH;
+
+      public class PathParameters {
+         public static final String REQUEST_ID = "requestId";
+         public static final String CONFIRMATION_ID = "confirmationId";
       }
    }
 
@@ -75,4 +90,41 @@ public abstract class KeyChangeTokenRequestsResource {
             httpHeaders,
             uriInfo);
    }
+
+   @POST
+   @Path(ConfirmKeyChange.RELATIVE_PATH)
+   @Consumes({ "application/json" })
+   @Produces({ "application/json" })
+   @ApiOperation(nickname = ConfirmKeyChange.CONFIRM_KEY_CHANGE, value = "Confirms that key change tokens have been provided to the customer.", notes = "Confirms that key change tokens have been provided successfully")
+   @ApiResponses(value = {
+         @ApiResponse(code = ConfirmKeyChange.SUCCESS, message = "Accepted", response = BasicAdvice.class),
+         @ApiResponse(code = 400, message = "Bad request", response = ErrorDetail.class),
+         @ApiResponse(code = 404, message = "Not found", response = ErrorDetail.class),
+         @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDetail.class),
+         @ApiResponse(code = 501, message = "Not implemented", response = ErrorDetail.class),
+         @ApiResponse(code = 503, message = "Service Unavailable", response = ErrorDetail.class),
+         @ApiResponse(code = 504, message = "Gateway Timeout", response = ErrorDetail.class) })
+   public void confirmKeyChange(
+         @ApiParam(value = "The randomly generated UUID of the original request.", required = true) @PathParam(ConfirmKeyChange.PathParameters.REQUEST_ID) String requestId,
+         @ApiParam(value = "The randomly generated UUID of this confirmation.", required = true) @PathParam(ConfirmKeyChange.PathParameters.CONFIRMATION_ID) String confirmationId,
+         @ApiParam(value = "A key change confirmation", required = true) KeyChangeConfirmation body,
+         @Context SecurityContext securityContext,
+         @Suspended AsyncResponse asyncResponse,
+         @Context Request request,
+         @Context HttpServletRequest httpServletRequest,
+         @Context HttpHeaders httpHeaders,
+         @Context UriInfo uriInfo) {
+
+      getResourceImplementation().confirmKeyChange(
+            requestId,
+            confirmationId,
+            body,
+            securityContext,
+            asyncResponse,
+            request,
+            httpServletRequest,
+            httpHeaders,
+            uriInfo);
+   }
+
 }