+++ src/test/java/io/electrum/prepaidutility/ModelTest.java
@@ -1,12 +1,12 @@
 package io.electrum.prepaidutility;
 
-import io.electrum.prepaidutility.model.ErrorDetail;
-import io.electrum.prepaidutility.model.Token;
-
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-public class ModelTests {
+import io.electrum.prepaidutility.model.ErrorDetail;
+import io.electrum.prepaidutility.model.Token;
+
+public class ModelTest {
 
    @Test
    public void testTokenTypeEnumOrdinals() {
@@ -27,7 +27,7 @@ public class ModelTests {
 
    @Test
    public void testErrorDetail_ErrorTypeEnumOrdinals() {
-      Assert.assertEquals(ErrorDetail.ErrorType.values().length, 21);
+      Assert.assertEquals(ErrorDetail.ErrorType.values().length, 22);
       Assert.assertEquals(ErrorDetail.ErrorType.DUPLICATE_RECORD.ordinal(), 0);
       Assert.assertEquals(ErrorDetail.ErrorType.FORMAT_ERROR.ordinal(), 1);
       Assert.assertEquals(ErrorDetail.ErrorType.FUNCTION_NOT_SUPPORTED.ordinal(), 2);
@@ -53,7 +53,7 @@ public class ModelTests {
 
    @Test
    public void testErrorDetail_RequestTypeEnumOrdinals() {
-      Assert.assertEquals(ErrorDetail.RequestType.values().length, 9);
+      Assert.assertEquals(ErrorDetail.RequestType.values().length, 10);
       Assert.assertEquals(ErrorDetail.RequestType.METER_LOOKUP_REQUEST.ordinal(), 0);
       Assert.assertEquals(ErrorDetail.RequestType.TOKEN_PURCHASE_REQUEST.ordinal(), 1);
       Assert.assertEquals(ErrorDetail.RequestType.TOKEN_PURCHASE_RETRY_REQUEST.ordinal(), 2);
@@ -63,5 +63,6 @@ public class ModelTests {
       Assert.assertEquals(ErrorDetail.RequestType.CONFIRMATION_ADVICE.ordinal(), 6);
       Assert.assertEquals(ErrorDetail.RequestType.REVERSAL_ADVICE.ordinal(), 7);
       Assert.assertEquals(ErrorDetail.RequestType.NOTIFY_TOKEN_PURCHASE.ordinal(), 8);
+      Assert.assertEquals(ErrorDetail.RequestType.TOKEN_PURCHASE_TRIAL_REQUEST.ordinal(), 9);
    }
 }